quick access (checkbox) to adjust reaction time...
Seems the bot my not be labeled too well. This controls the speed of the bot. The reaction time. The wait timers. Everything. If you want to 'power level' this is what you click. The bot will do what it does without hinderance (though some classes may not have adequate conditions for spells). This has always been here, just hidden in the menu.