more quests added. level 15+ both factions
a little time consuming because I am going through them pretty much 1 by 1 at picking the ones that are the best for the bot - or ordering them for the bot to make a nav path without running through a bunch of mobs that will kill it. doesn't take long to add quests to the DB. trial and error. levels 1-15 should cause no problems whatsoever while running it.