bug fixes + quests added
The nav of the bot isn't the greatest and the quester is losing its ability to find paths to quest targets inside buildings. Trying to add the ability to get close enough to the quest person and force the bot to move in their general direction without navigation. This problem was specifically seen doing redridge quests.