no more nav crashes!
unstuck script, along with all the calls to the nav were causing the problem. mostly the unstuck script being outside of the grind script wait timer and all the calls were corrupting the nav, it seems. both unstuck functions work as intended with their own separate wait timers now. if one unstuck script doesn't activate then the other will.

first unstuck checks for movement of current players xyz based on time and distance and moves 3 or more different directions if xyz doesn't change enough in long enough time. since changing in z axis causes large jumps in 3d distance (in yards), the original unstuck must still be preserved to knock the bot in different directions when needed.

I tested this for several hours in a spot that you cannot move or jump out of and let the bot run swinging nav paths back and forth, reloading the bot several times and reloading the game several times to make sure it wasn't a fluke.