hotspot logic changes
bot will now be at its hotspot if the distance to hotspot is within distance to grind from the hotspot. - i.e. the bot is 300 yards away from hotspot and the bot is allowed to travel 300 yards away from the hotspot, it will initiate combat at that 300 yard mark.

this is combined with the targeting functions which account for target distance from hotspot and player distance to target. if the target is within distance to hotspot and within range of bot then it will become the bots target.

autopath nodes should take over and if the bot leaves the hotspot area to vendor or other reason then it will return directly to the hotspot area and pick up on auto path nodes until a target is found

if you MANUALLY change distance to hotspot while the bot has been grinding and the bot has made auto path nodes it will currently still walk those nodes and not reset them, or find a target outside of hotspot distance even if nodes are made. if the bot is manually being controlled at that point then a button can be added to reset saved nav points