minor changes
found a good spot to insert code for hotspots. changing status messages and some minor changes to targeting and group behavior. 

trying to open the 'walk paths' menu for use so paths can be added for a new script strictly for gathering professions. make a path, walk a path, gather on path, kill target if in combat, walk path, gather on path.